+++ git-index/src/file.rs
@@ -1,14 +1,16 @@
 pub mod init {
     #![allow(unused)]
 
-    use crate::file::decode;
-    use crate::{File, State};
-    use memmap2::Mmap;
     use std::path::{Path, PathBuf};
 
+    use memmap2::Mmap;
+
+    use crate::{extension, file::header, File, State};
+
     mod error {
-        use crate::file::decode;
         use quick_error::quick_error;
+
+        use crate::file::header;
 
         quick_error! {
             #[derive(Debug)]
@@ -18,7 +20,7 @@ pub mod init {
                     source(err)
                     from()
                 }
-                DecodeHeader(err: decode::header::Error) {
+                DecodeHeader(err: header::decode::Error) {
                     display("The header could not be understood")
                     source(err)
                     from()
@@ -39,8 +41,8 @@ pub mod init {
                 (data, filetime::FileTime::from_last_modification_time(&file.metadata()?))
             };
 
-            let (version, num_entries, post_header_data) = decode::header(&data, object_hash)?;
-            let start_of_extensions = decode::extension::end_of_index_entry(&data, object_hash);
+            let (version, num_entries, post_header_data) = header::decode(&data, object_hash)?;
+            let start_of_extensions = extension::EndOfIndexEntry::from_bytes(&data, object_hash);
 
             Ok(File {
                 state: State { timestamp: mtime },
@@ -50,70 +52,32 @@ pub mod init {
     }
 }
 
-pub mod decode {
-    use crate::Version;
+pub mod header {
+    pub(crate) const SIZE: usize = 4 /*signature*/ + 4 /*version*/ + 4 /* num entries */;
 
-    fn extension(data: &[u8]) -> ([u8; 4], u32, &[u8]) {
-        let (signature, data) = data.split_at(4);
-        let (size, data) = data.split_at(4);
-        (signature.try_into().unwrap(), read_u32(size), data)
-    }
+    pub mod decode {
+        use quick_error::quick_error;
 
-    pub(crate) mod extension {
-        use crate::extension::EndOfIndexEntry;
-        use crate::file::decode;
-        use crate::file::decode::read_u32;
-
-        pub fn end_of_index_entry(data: &[u8], object_hash: git_hash::Kind) -> Option<EndOfIndexEntry> {
-            let hash_len = object_hash.len_in_bytes();
-            if data.len() < EndOfIndexEntry::SIZE_WITH_HEADER + hash_len {
-                return None;
-            }
-
-            let start_of_eoie = data.len() - EndOfIndexEntry::SIZE_WITH_HEADER - hash_len;
-            let data = &data[start_of_eoie..][..hash_len];
-
-            let (signature, ext_size, data) = decode::extension(data);
-            if &signature != EndOfIndexEntry::SIGNATURE || ext_size as usize != EndOfIndexEntry::SIZE {
-                return None;
-            }
-
-            let (offset, hash) = data.split_at(4);
-            let offset = read_u32(offset) as usize;
-            if offset < decode::header::SIZE {
-                return None;
-            }
-            todo!("eoie")
-        }
-    }
-
-    pub mod header {
-        pub(crate) const SIZE: usize = 4 /*signature*/ + 4 /*version*/ + 4 /* num entries */;
-
-        mod error {
-            use quick_error::quick_error;
-
-            quick_error! {
-                #[derive(Debug)]
-                pub enum Error {
-                    Corrupt(message: &'static str) {
-                        display("{}", message)
-                    }
-                    UnsupportedVersion(version: u32) {
-                        display("Index version {} is not supported", version)
-                    }
+        quick_error! {
+            #[derive(Debug)]
+            pub enum Error {
+                Corrupt(message: &'static str) {
+                    display("{}", message)
+                }
+                UnsupportedVersion(version: u32) {
+                    display("Index version {} is not supported", version)
                 }
             }
         }
-        pub use error::Error;
     }
+    use crate::{util::read_u32, Version};
 
-    pub(crate) fn header(
+    pub(crate) fn decode(
         data: &[u8],
         object_hash: git_hash::Kind,
-    ) -> Result<(crate::Version, u32, &[u8]), header::Error> {
+    ) -> Result<(crate::Version, u32, &[u8]), decode::Error> {
         if data.len() < (3 * 4) + object_hash.len_in_bytes() {
-            return Err(header::Error::Corrupt(
+            return Err(decode::Error::Corrupt(
                 "File is too small even for header with zero entries and smallest hash",
             ));
         }
@@ -121,7 +85,7 @@ pub mod decode {
         const SIGNATURE: &[u8] = b"DIRC";
         let (signature, data) = data.split_at(4);
         if signature != SIGNATURE {
-            return Err(header::Error::Corrupt(
+            return Err(decode::Error::Corrupt(
                 "Signature mismatch - this doesn't claim to be a header file",
             ));
         }
@@ -131,16 +95,11 @@ pub mod decode {
             2 => Version::V2,
             3 => Version::V3,
             4 => Version::V4,
-            unknown => return Err(header::Error::UnsupportedVersion(unknown)),
+            unknown => return Err(decode::Error::UnsupportedVersion(unknown)),
         };
         let (entries, data) = data.split_at(4);
         let entries = read_u32(entries);
 
         Ok((version, entries, data))
-    }
-
-    #[inline]
-    fn read_u32(b: &[u8]) -> u32 {
-        u32::from_be_bytes(b.try_into().unwrap())
     }
 }
